test(select): migrate color
