+++ core/src/components/select/test/color/select.e2e.ts-snapshots/select-outline-color-md-ltr-Mobile-Chrome-linux.png
