+++ core/src/components/select/test/color/select.e2e.ts-snapshots/select-outline-color-md-ltr-Mobile-Firefox-linux.png
