+++ core/src/components/select/test/color/select.e2e.ts-snapshots/select-solid-color-md-ltr-Mobile-Safari-linux.png
